+++ .gitignore
@@ -1,2 +1,4 @@
 information_rl.csv
 __pycache__/
+
+data/XuetangE
