+++ table_ingest.py
@@ -1,5 +1,5 @@
 import sketches
-from data_prep import Table
+from table import Table
 import pandas as pd
 
 
@@ -52,3 +52,104 @@ core = ["data/XuetangE/train_log.csv", "data/XuetangE/train_truth.csv"]
 keys = ["user_id", "course_id"]
 main(core, 'enroll_id', paths, keys, 'truth')
 print('ala')
+
+
+
+def data_preprocessing(data_sources, y):
+    """_summary_
+
+    Returns:
+        tables: {id: [dataframe, path_cost, best_path]}
+        best_path for table i: [(a, feat_a), (b, feat_c)]
+    """
+    features = {}  # feature_name: [table_id]
+    
+    target_id = -1  # id of the table that contains y
+    # TODO: account for multiple tables having the target_id
+    
+    table_id = 1
+    id_to_table = {}  # id: dataframe
+    for location in data_sources:  # loading data and creating table ids
+        df = pd.read_csv(location)  # load into dataframe
+        # create a lookup of which features exist in which table
+        for col in df.columns:
+            if col in features:
+                features[col].append(table_id)
+            else:
+                features[col] = [table_id]
+            if col == y:
+                target_id = table_id
+        id_to_table[table_id] = df  # store dataframe based on table id
+        table_id += 1
+    
+    num_tables = table_id  # len(id_to_table)
+    table_edges = []
+    for i in range(num_tables):  # initialize edges
+        table_edges.append([np.inf]*num_tables)
+    # populating edges with cost of join
+    for feature_name, table_ids in features.items():
+        # identify joinable like elements and set as joinable
+        if len(table_ids) < 1:
+            pass
+        for tid1 in table_ids:
+            for tid2 in table_ids:
+                if tid1 == tid2:
+                    table_edges[tid1, tid2] = 0
+                else:
+                    table_edges[tid1, tid2] = calc_join_cost(id_to_table[tid1], id_to_table[tid2], feature_name)
+    
+    # calculating best join plan
+    optimized_path, optimized_path_cost =  shortest_path(id_to_table.keys(), table_edges)
+    
+    # get optimized path from every table to the table with the target feature
+    
+    tables = {}
+    # store path with cost for each table
+    for starting_table in range(num_tables):
+        # for each starting table, save the cost of joining to the target_id table
+        # aka the best path to the table with the target variable
+        info = [id_to_table[starting_table]]
+        info.append(optimized_path_cost[starting_table][target_id])
+        best_path = compute_best_path(optimized_path, starting_table, target_id)
+        info.append(best_path)
+
+    return tables
+
+def calc_join_cost(t1, t2, feature):
+    """ does a rough estimation of the cost of joining tables, t1 and t2 via feature
+
+    Args:
+        t1 (dataframe): _description_
+        t2 (dataframe): _description_
+        feature (str): _description_
+    """
+    ...
+    return 0
+
+
+def shortest_path(edges, verticies):
+    """ uses Floyd-Warshall to generate the shortest path from every node to every other node
+
+    Args:
+        edges (_type_): _description_
+        verticies (_type_): _description_
+            
+    Returns:
+        path: _description_
+        cost: _description_
+    """
+    
+def compute_best_path(optimized_path, start_d, end_id):
+    """ takes in the optimized path table and returns the sequence of 
+        features and tables that are needed to join the two table ids
+
+    Args:
+        optimized_path (_type_): _description_
+        start_d (_type_): _description_
+        end_id (_type_): _description_
+
+    Returns:
+        _type_: _description_
+    """
+    path = []
+    return path
